# backend/docs/swagger.yaml
definitions:
...
  dto.ProductSupplyDTO:
    properties:
      delivery_date:
        type: string
      price:
        type: number
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: number
      supplier_name:
        type: string
      supply_id:
        type: string
    type: object
  dto.ProductTypeDTO:
    properties:
      for_adult:
        type: boolean
      name:
        type: string
      type_id:
        type: string
    type: object
...
  models.Product:
    properties:
      manufacturer:
        type: string
      name:
        type: string
      product_id:
        type: string
      type:
        $ref: '#/definitions/models.ProductType'
      type_id:
        type: string
      volume:
        type: number
      weight:
        type: number
    type: object
...
  models.ProductSupply:
    properties:
      delivery_date:
        type: string
      price:
        type: number
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: number
      supplier:
        $ref: '#/definitions/models.Supplier'
      supplier_id:
        type: string
      supply_id:
        type: string
    type: object
  models.ProductType:
    properties:
      for_adult:
        type: boolean
      name:
        type: string
      type_id:
        type: string
    type: object
...
  models.Supplier:
    properties:
      email:
        type: string
      inn:
        type: string
      kpp:
        type: string
      name:
        type: string
      ogrn:
        type: string
      phone:
        type: string
      supplier_id:
        type: string
    type: object
  response.Response:
    properties:
      error:
        description: Error if response is error
        type: string
      status:
        description: Response status
        type: string
    type: object
info:
  contact: {}
paths:
...
  /product-supplies:
    get:
      description: Returns a list of product supplies filtered by product, supplier
        and date range.
      parameters:
      - description: Product ID
        format: uuid
        in: query
        name: product_id
        type: string
      - description: Supplier ID
        format: uuid
        in: query
        name: supplier_id
        type: string
      - description: Start date (RFC3339)
        in: query
        name: date_from
        type: string
      - description: End date (RFC3339)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductSupplyDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get product supplies
      tags:
      - product-supplies
    post:
      consumes:
      - application/json
      description: Creates a new product supply.
      parameters:
      - description: Product supply
        in: body
        name: productSupply
        required: true
        schema:
          $ref: '#/definitions/models.ProductSupply'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductSupply'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create product supply
      tags:
      - product-supplies
  /product-supplies/{id}:
    delete:
      description: Deletes a product supply by ID.
      parameters:
      - description: Product supply ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete product supply
      tags:
      - product-supplies
    put:
      consumes:
      - application/json
      description: Updates an existing product supply.
      parameters:
      - description: Product supply ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Product supply
        in: body
        name: productSupply
        required: true
        schema:
          $ref: '#/definitions/models.ProductSupply'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update product supply
      tags:
      - product-supplies
  /product-types:
    get:
      description: Returns a list of product types filtered by search string and adult
        flag.
      parameters:
      - description: Search string
        in: query
        name: search
        type: string
      - description: Filter by adult products
        in: query
        name: for_adult
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductTypeDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get product types
      tags:
      - product-types
    post:
      consumes:
      - application/json
      description: Creates a new product type.
      parameters:
      - description: Product type
        in: body
        name: productType
        required: true
        schema:
          $ref: '#/definitions/models.ProductType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create product type
      tags:
      - product-types
  /product-types/{id}:
    delete:
      description: Deletes a product type by ID.
      parameters:
      - description: Product type ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete product type
      tags:
      - product-types
    put:
      consumes:
      - application/json
      description: Updates an existing product type.
      parameters:
      - description: Product type ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Product type
        in: body
        name: productType
        required: true
        schema:
          $ref: '#/definitions/models.ProductType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update product type
      tags:
      - product-types
...
swagger: "2.0"